feature SummaryTable: create summary table and modify chip to display it
