+++ src/components/Chats.tsx
@@ -1,6 +1,7 @@
 import React, {useState, useEffect} from "react";
 import ChatTable from "./ChatTable";
 import ChatTableUserB from "./ChatTableUserB";
+import SummaryTable from "./SummaryTable";
 import Login from "./Login";
 import {useUserContext} from "../utils/utils";
 import {List, Chip} from "@telegram-apps/telegram-ui";
@@ -9,6 +10,7 @@ const Chats: React.FC<{backendUrl: string}> = ({backendUrl}) => {
   const {user, isLoggedIn} = useUserContext();
   const [showChatTable, setShowChatTable] = useState<boolean>(false);
   const [showChatTableUserB, setShowChatTableUserB] = useState<boolean>(false);
+  const [showSummaryTable, setShowSummaryTable] = useState<boolean>(false);
   const [showLogin, setShowLogin] = useState<boolean>(false);
 
   // Update state based on user profile and chats
@@ -21,12 +23,14 @@ const Chats: React.FC<{backendUrl: string}> = ({backendUrl}) => {
           );
           setShowChatTableUserB(true);
           setShowChatTable(false);
+          setShowSummaryTable(false);
         } else {
           console.log(
             "User doesn't have a profile and has no chats, showing Login",
           );
           setShowChatTable(false);
           setShowChatTableUserB(false);
+          setShowSummaryTable(false);
           setShowLogin(true);
           console.log(
             "User doesn't have a profile and has no chats, showing Login",
@@ -37,9 +41,11 @@ const Chats: React.FC<{backendUrl: string}> = ({backendUrl}) => {
           console.log("User has a profile, showing ChatTable");
           setShowChatTable(true);
           setShowChatTableUserB(false);
+          setShowSummaryTable(true);
         } else {
           setShowChatTable(false);
           setShowChatTableUserB(false);
+          setShowSummaryTable(false);
           setShowLogin(true);
         }
         setShowChatTableUserB(false);
@@ -51,6 +57,7 @@ const Chats: React.FC<{backendUrl: string}> = ({backendUrl}) => {
     setShowLogin(false);
     setShowChatTable(true);
     setShowChatTableUserB(false);
+    setShowSummaryTable(true);
   };
 
   const handleMyChatsClick = () => {
@@ -59,6 +66,7 @@ const Chats: React.FC<{backendUrl: string}> = ({backendUrl}) => {
       console.log("User is logged in");
       setShowChatTable(true);
       setShowChatTableUserB(false);
+      setShowSummaryTable(false);
     } else {
       console.log("User is not logged in");
       setShowLogin(true);
@@ -73,36 +81,54 @@ const Chats: React.FC<{backendUrl: string}> = ({backendUrl}) => {
     } else {
       setShowChatTableUserB(true);
       setShowChatTable(false);
+      setShowSummaryTable(false);
       setShowLogin(false);
     }
   };
 
+  const handleMySummaryClick = () => {
+    console.log("My Summary clicked");
+    if (isLoggedIn) {
+      setShowSummaryTable(true);
+      setShowChatTable(false);
+      setShowLogin(false);
+      setShowChatTableUserB(false);
+    } else {
+      setShowLogin(true);
+    }
+  };
+
   return (
-    <div className='p-5 max-w-xl mx-auto text-center'>
+    <div className='p-2 max-w-xl mx-auto text-center'>
       <List
-        className='p-5 rounded-lg shadow mb-4 '
+        className='p-2 rounded-lg shadow mb-4 overflow-auto w-fit'
         style={{background: "var(--tgui--secondary_bg_color)"}}
       >
-        <div className='flex gap-4'>
+        <div className='flex gap-2'>
           <Chip
             mode={showChatTable ? "elevated" : "mono"}
-            after={<span className='chip-icon'>👉</span>}
             onClick={handleMyChatsClick}
           >
-            My chats
+            My chats 💬
           </Chip>
           <Chip
             mode={showChatTableUserB ? "elevated" : "mono"}
-            after={<span className='chip-icon'>📧</span>}
             onClick={handleMyInvitationsClick}
           >
-            My invitations
+            My invitations 📩
+          </Chip>
+          <Chip
+            mode={showSummaryTable ? "elevated" : "mono"}
+            onClick={handleMySummaryClick}
+          >
+            My Summary 📝
           </Chip>
         </div>
       </List>
       <div className='w-full'>
         {showChatTable && <ChatTable user={user} backendUrl={backendUrl} />}
         {showChatTableUserB && <ChatTableUserB backendUrl={backendUrl} />}
+        {showSummaryTable && <SummaryTable user={user} />}
       </div>
       {showLogin && (
         <Login onLoginSuccess={handleLoginSuccess} backendUrl={backendUrl} />
